Adds Open Simplex noise library.
Depends on https://github.com/deerel/OpenSimplexNoise
It may not be the fastest computationally, but it was the fastest to get up and running, I am planning to add more noise components in the future, when I get a successful compile using another library.
I was considering FastNoise2, but I was having trouble compiling it using MinGW and hopsan.

This commit adds an initial 3 blocks:
 - 1D simplex noise, based on simulation time with various constants to tweak to requirements.
 - Variable 1D Simplex based on an explicit x input. Fewer constants to tweak.
 - A Variable 4D Simplex block with X, Y, Z, W inputs and a single output signal.

 No Help files yet.
 1 Example